+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Snap to the target in `SmoothNudge` when the current value is close enough.
+- Use `debug!` instead of `trace!` for events logging.
 
 ## [0.6.0] - 2025-01-13